# docs/swagger.yaml
basePath: /
definitions:
  go_auth_internal_apierror.Body:
    properties:
      error:
        $ref: '#/definitions/go_auth_internal_apierror.Detail'
    type: object
  go_auth_internal_apierror.Detail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  go_auth_internal_permission.Permission:
    properties:
      active:
        type: boolean
      application_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  internal_role.AssignPermissionRequest:
    properties:
      permission_id:
        type: integer
    required:
    - permission_id
    type: object
  internal_role.CreateRequest:
    properties:
      application_id:
        type: integer
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - application_id
    - name
    - slug
    type: object
  internal_role.Role:
    properties:
      active:
        type: boolean
      application_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      uuid:
        type: string
    type: object
  internal_role.UpdateRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  internal_user.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
      uuid:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API de autenticação e autorização em Go com JWT, suportando múltiplas
    aplicações
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Auth API
  version: "1.0"
paths:
...
  /api/admin/roles:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all roles, optionally filtered by application_id
      parameters:
      - description: Filter by application ID
        in: query
        name: application_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_role.Role'
            type: array
      security:
      - Bearer: []
      summary: Get all roles
      tags:
      - Admin - Roles
    post:
      consumes:
      - application/json
      description: Create a new role associated with an application
      parameters:
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/internal_role.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_role.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Create a new role
      tags:
      - Admin - Roles
  /api/admin/roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Delete a role
      tags:
      - Admin - Roles
    get:
      consumes:
      - application/json
      description: Retrieve a specific role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_role.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get role by ID
      tags:
      - Admin - Roles
    put:
      consumes:
      - application/json
      description: Update an existing role's fields
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/internal_role.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Update a role
      tags:
      - Admin - Roles
  /api/admin/roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Retrieve all permissions assigned to a specific role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go_auth_internal_permission.Permission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: List role permissions
      tags:
      - Admin - Role-Permission Relationships
    post:
      consumes:
      - application/json
      description: Assign a permission to a specific role (both must belong to the
        same application)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/internal_role.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Assign permission to role
      tags:
      - Admin - Role-Permission Relationships
  /api/admin/roles/{id}/permissions/{permissionId}:
    delete:
      consumes:
      - application/json
      description: Remove a permission from a specific role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Remove permission from role
      tags:
      - Admin - Role-Permission Relationships
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users registered in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_user.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: List all users
      tags:
      - Admin - Users
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username, email and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "409":
          description: Email or username already taken
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      summary: Register a new user
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Digite "Bearer " seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"